+++ unique_toolkit/CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), 
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.36] - 2024-11-19
+- Add possibility to return the response from the download file from chat request
+- Add possibility to not specify a filename and use filename from response headers
+
+## [0.5.35] - 2024-11-18
+- Add the possibilty to upload files without triggering ingestion by setting `skip_ingestion` to `True` in `ContentService.upload_content`
+
+## [0.5.34] - 2024-11-15
+- Add `content_id_to_translate` to `EventAdditionalParameters`
+
 ## [0.5.33] - 2024-10-30
 - Force randomizing tool_call_id. This is helpful to better identify the tool_calls.